NetworkManager 구성 - TMDB Request enum 정리 및 케이스 구분 - MainView 오늘의 영화 항목 네트워크 통신 연결

+++ MyMovieBox/Model/Movie(Trending, Search).swift
@@ -15,7 +15,7 @@ struct Movie: Decodable {
     let backdropPath: String
     let releaseDate: String
     let genreIds: [Int]
-    let voteAverage: Int
+    let voteAverage: Double
     
     enum CodingKeys: String, CodingKey {
         case id
